+++ hugr-model/src/v0/mod.rs
@@ -281,6 +281,8 @@ pub mod binary;
 pub mod scope;
 pub mod text;
 
+pub use bumpalo;
+
 macro_rules! define_index {
     ($(#[$meta:meta])* $vis:vis struct $name:ident(pub u32);) => {
         #[repr(transparent)]